web: Move PanicLinkInfo and createErrorFooter to own file
